refactor: split excuse.go into excuse_entry.go and excuse_template.go
