+++ internal/models/excuse_entry.go
@@ -4,7 +4,6 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/lib/pq"
 )
 
 type ExcuseEntry struct {
@@ -16,14 +15,4 @@ type ExcuseEntry struct {
 	TemplateID *string   `gorm:"size:255"`
 	CreatedAt  time.Time `gorm:"default:CURRENT_TIMESTAMP"`
 	UpdatedAt  time.Time `gorm:"default:CURRENT_TIMESTAMP"`
-}
-
-type ExcuseTemplate struct {
-	ID        string         `gorm:"primaryKey;size:255"` // "gravity-strong" etc
-	Text      string         `gorm:"type:text;not null"`
-	PackID    string         `gorm:"size:255;default:'core'"` // "core", "pack.surreal", etc
-	IsActive  bool           `gorm:"default:true"`
-	IsPremium bool           `gorm:"default:false"`
-	Tags      pq.StringArray `gorm:"type:text[]"`
-	CreatedAt time.Time      `gorm:"default:CURRENT_TIMESTAMP"`
 }
